Update disturbance_model.m
Changed the second model: now the 2nd linear filter has 4 poles instead of two, to get realizations that are 2 times differentiable. In the parts where conditional means are computed, we need to invert matrices. This part needs to be checked, in particular when p_w is large and dt are small in order to make sure that the matrices are invertible.