Added option to pass rng object to noise_interpolation, as well as fixed minor bug with noise interpolation
Added option to pass rng object to noise_interpolation ensuring that the cost function always gives same value for same parameters.
Fixed bug with noise interpolation that meant that if w(t) was computed twice for the same value of t, it would return two different values, despite the previous value of w(t) having been stored. Basically, the previous value was not correctly conditioned on if a t was requested that was exactly equal to a previously requested t.


Former-commit-id: 4601b1bafedc7dccb9253ad09f2bb2b0ea5a501d [formerly 8b5179ffe35095a74daa3596a182df3168b68cbd]
Former-commit-id: 14e28d046b05888ab4d6e09c2355fd71f173de23